Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/intrinsicsize/intrinsicsize-with-responsive-images.tentative-expected.txt



FAIL Test image /images/green.svg with no specified sizes, width, or height assert_equals: width expected 400 but got 33
FAIL Test image /images/green.svg with width = 800, no specified sizes, or height assert_equals: height expected 1000 but got 400
FAIL Test image /images/green.svg with width = 800, height = 800, and no specified sizes assert_equals: naturalWidth expected 400 but got 800
FAIL Test image /images/green.png with no specified sizes, width, or height assert_equals: width expected 400 but got 33
FAIL Test image /images/green.png with width = 800, no specified sizes, or height assert_equals: height expected 1000 but got 400
FAIL Test image /images/green.png with width = 800, height = 800, and no specified sizes assert_equals: naturalWidth expected 400 but got 100
FAIL Test image (32 x 32) with sizes = 100 and srcset descriptor = 32w assert_equals: height expected 125 but got 100
